Some fixes, added documentation
