+++ docs/useRebar/document/document-account.md
@@ -154,151 +154,13 @@ function someFunction(player: alt.Player) {
 }
 ```
 
-### permissions
+### [permissions](/userebar/systems/permissions/playerPermissions.md#useaccount)
 
-You can grant/revoke permissions to the account and check if account has them.
+Click on the link above to see how to use account permissions.
 
-#### addPermission
+### [groups](/userebar/systems/permissions/playerGroups.md#useaccount)
 
-Grants a permission to the account.
-
-```ts
-async function someFunction(player: alt.Player) {
-    const document = Rebar.document.account.useAccount(player);
-    await document.permissions.addPermission('admin');
-}
-```
-
-#### removePermission
-
-Revokes permission from the account.
-
-```ts
-async function someFunction(player: alt.Player) {
-    const document = Rebar.document.account.useAccount(player);
-    await document.permissions.removePermission('admin');
-}
-```
-
-#### hasPermission
-
-Checks if account has a permission.
-
-```ts
-function someFunction(player: alt.Player) {
-    const document = Rebar.document.account.useAccount(player);
-    const hasPerm: boolean = document.permissions.hasPermission('admin');
-}
-```
-
-#### hasAllPermissions
-
-Checks if account has all of the provided permissions:
-
-```ts
-function someFunction(player: alt.Player) {
-    const document = Rebar.document.account.useAccount(player);
-    const hasAllPerm: boolean = document.permissions.hasAllPermissions(['admin', 'support']);
-}
-```
-
-#### hasAnyPermission
-
-Checks if account has at least one of the provided permissions:
-
-```ts
-function someFunction(player: alt.Player) {
-    const document = Rebar.document.account.useAccount(player);
-    const hasAnyPerm: boolean = document.permissions.hasAnyPermission(['admin', 'support']);
-}
-```
-
-### groupPermissions
-
-Permission groups allow you to assign permissions under a specific group name for an account.
-
-#### addPermissions
-
-Adds permission to specified group. If an account had no group before - it will create it.
-
-```ts
-async function someFunction(player: alt.Player) {
-    const document = Rebar.document.account.useAccount(player);
-    await document.groupPermissions.addPermissions('admin', 'noclip');
-}
-```
-
-#### removePermissions
-
-Removes permission from group, it will also remove a group, if there are no permissions left.
-
-```ts
-async function someFunction(player: alt.Player) {
-    const document = Rebar.document.account.useAccount(player);
-    await document.groupPermissions.removePermissions('admin', 'noclip');
-}
-```
-
-#### removeGroup
-
-Completely removes group from an account with all its permissions.
-
-```ts
-async function someFunction(player: alt.Player) {
-    const document = Rebar.document.account.useAccount(player);
-    await document.groupPermissions.removeGroup('admin');
-}
-```
-
-#### hasGroup
-
-Checks if account belongs to group.
-
-```ts
-function someFunction(player: alt.Player) {
-    const document = Rebar.document.account.useAccount(player);
-    const isAdmin: boolean = document.groupPermissions.hasGroup('admin');
-}
-```
-
-#### hasGroupPerm
-
-Checks if account belongs to group and has a specific permission.
-
-```ts
-function someFunction(player: alt.Player) {
-    const document = Rebar.document.account.useAccount(player);
-    const canNoclip: boolean = document.groupPermissions.hasGroupPerm('admin', 'noclip');
-}
-```
-
-#### hasAtLeastOneGroupPerm
-
-Checks if account belongs to group and has any of provided permissions.
-
-```ts
-function someFunction(player: alt.Player) {
-    const document = Rebar.document.account.useAccount(player);
-    const canBanOrNoclip: boolean = document.groupPermissions.hasAtLeastOneGroupPerm('admin', ['noclip', 'ban']);
-}
-```
-
-#### hasAtLeastOneGroupWithSpecificPerm
-
-Checks if account belongs to any of the group and has corresponding rights.
-
-```ts
-function someFunction(player: alt.Player) {
-    const document = Rebar.document.account.useAccount(player);
-    const isStaff = document.groupPermissions.hasAtLeastOneGroupWithSpecificPerm({
-        admin: ['noclip', 'ban'],
-        support: ['answerReport'],
-    });
-    // This will be true if:
-    // 1. Account has `admin` group with `noclip` OR/AND `ban` permission.
-    // 2. Account has `support` group with `answerReport` permission.
-}
-```
+Click on the link above to see how to use account groups.
 
 ## useAccountEvents
