+++ docs/useRebar/document/document-virtual.md
@@ -4,6 +4,8 @@ See [Documents Section](./index.md) for further information on documents.
 
 ## useVirtual
 
+If document is not found in the database, it will return `undefined`.
+
 ```ts
 import { useRebar } from '@Server/index.js';
 
@@ -15,7 +17,7 @@ interface CustomDocument {
     test2: string;
 }
 
-const virtualDocument = Rebar.document.virtual.useVirtual<CustomDocument>('663ce39eb270106cf02fb7e3', 'SomeCollection');
+const virtualDocument = await Rebar.document.virtual.useVirtual<CustomDocument>('663ce39eb270106cf02fb7e3', 'SomeCollection');
 ```
 
 ### get
@@ -24,11 +26,14 @@ Get all data for a document
 
 ```ts
 async function someFunction() {
-    const virtualDocument = Rebar.document.virtual.useVirtual<CustomDocument>(
+    const virtualDocument = await Rebar.document.virtual.useVirtual<CustomDocument>(
         '663ce39eb270106cf02fb7e3',
         'SomeCollection',
     );
-    const data = await virtualDocument.get();
+    if (!virtualDocument) {
+        return;
+    }
+    const data = virtualDocument.get();
     console.log(data);
 }
 ```
@@ -39,11 +44,14 @@ Get a specific field from a document
 
 ```ts
 async function someFunction() {
-    const virtualDocument = Rebar.document.virtual.useVirtual<CustomDocument>(
+    const virtualDocument = await Rebar.document.virtual.useVirtual<CustomDocument>(
         '663ce39eb270106cf02fb7e3',
         'SomeCollection',
     );
-    const test1 = await virtualDocument.getField('test1');
+    if (!virtualDocument) {
+        return;
+    }
+    const test1 = virtualDocument.getField('test1');
     console.log(test1);
 }
 ```
@@ -54,11 +62,13 @@ Set a specific field for the document, and save to the database
 
 ```ts
 async function someFunction() {
-    const virtualDocument = Rebar.document.virtual.useVirtual<CustomDocument>(
+    const virtualDocument = await Rebar.document.virtual.useVirtual<CustomDocument>(
         '663ce39eb270106cf02fb7e3',
         'SomeCollection',
     );
-
+    if (!virtualDocument) {
+        return;
+    }
     await virtualDocument.set('test1', 'hi');
 }
 ```
@@ -69,11 +79,46 @@ Set multiple fields for the document, and save to the database
 
 ```ts
 async function someFunction() {
-    const virtualDocument = Rebar.document.virtual.useVirtual<CustomDocument>(
+    const virtualDocument = await Rebar.document.virtual.useVirtual<CustomDocument>(
         '663ce39eb270106cf02fb7e3',
         'SomeCollection',
     );
-
+    if (!virtualDocument) {
+        return;
+    }
     await virtualDocument.setBulk({ test1: 'hi', test2: 'hi' });
 }
 ```
+
+### refresh
+
+Refresh the document from the database
+
+```ts
+async function someFunction() {
+    const virtualDocument = await Rebar.document.virtual.useVirtual<CustomDocument>(
+        '663ce39eb270106cf02fb7e3',
+        'SomeCollection',
+    );
+    if (!virtualDocument) {
+        return;
+    }
+    await virtualDocument.refresh();
+}
+```
+
+### [permissions](/userebar/systems/permissions/playerPermissions.md#usevirtual)
+
+If you want to modify permissions for a document, you can use the `permissions` property.
+
+It is available only for `Accounts` and `Characters` collections.
+
+Click the link above(heading) to see more details on how to use this property.
+
+### [groups](/userebar/systems/permissions/playerGroups.md#usevirtual)
+
+If you want to modify groups for a document, you can use the `groups` property.
+
+It is available only for `Accounts` and `Characters` collections.
+
+Click the link above(heading) to see more details on how to use this property.
