+++ docs/useRebar/systems/permissions/index.yml
@@ -1,0 +1,3 @@
+expanded: false
+order: 550
+label: 'permissions'
