+++ docs/useRebar/systems/useMessenger.md
@@ -4,6 +4,9 @@ The messenger system allows for developers to easily send up messages from playe
 
 However, the messages are not automatically sent to other players. You as a developer get to decide who sees what messages or if they see text messages at all.
 
+Commands have permission system built-in, so you can easily allow players to use certain commands based on their permissions.
+For more details on how to use permissions, check [useEntityPermissions](/userebar/systems/permissions/useEntityPermissions.md).
+
 ## Usage
 
 ```ts
@@ -26,7 +29,7 @@ messenger.commands.register({
 messenger.commands.register({
     name: '/goto',
     desc: '/goto [x][y][z]',
-    options: { accountPermissions: ['moderator'] },
+    options: { permissions: ['moderator'] },
     callback: (player: alt.Player, x: string, y: string, z: string) => {
         try {
             const pos = new alt.Vector3(parseFloat(x), parseFloat(y), parseFloat(z));
