+++ src/main/server/utility/protectCallback.ts
@@ -1,24 +1,18 @@
 import * as alt from 'alt-server';
 import { usePlayer } from '../player/index.js';
-import {PermissionOptions} from "@Shared/types/index.js";
-import {useEntityPermission} from "@Server/systems/permissions/entityPermissions.js";
+import { PermissionOptions } from '@Shared/types/index.js';
+import { useEntityPermissions } from '@Server/systems/permissions/entityPermissions.js';
 
-function hasPermission(
-    player: alt.Player,
-    permissionOptions: PermissionOptions,
-) {
+function hasPermission(player: alt.Player, permissionOptions: PermissionOptions) {
     const rPlayer = usePlayer(player);
     if (!rPlayer.isValid()) {
         return false;
     }
 
-    return useEntityPermission(permissionOptions).check(player, true);
+    return useEntityPermissions(permissionOptions).check(player);
 }
 
-export function useProtectCallback(
-    callback: Function,
-    permissionOptions: PermissionOptions,
-) {
+export function useProtectCallback(callback: Function, permissionOptions: PermissionOptions) {
     return (player: alt.Player, ...args: any[]) => {
         if (!hasPermission(player, permissionOptions)) {
             return;
